Decompose three god objects in @grackle-ai/core (#1474)
credential-bundle.ts (722 lines) → 5 focused modules behind a
re-export barrel: credential-types (shared interfaces), credential-needs
(pure RFC 9728 descriptor derivation), credential-inspector (pure expiry
inspection), credential-preflight (pure cross-check), credential-materializer
(side-effectful fs/env/gh-CLI bundle builder).

stream-registry.ts (857 lines) → 4 modules + shared-state substrate:
stream-registry-state (all Maps, AsyncQueue, private helpers),
stream-storage (create/get/delete/list), subscription-manager
(subscribe/unsubscribe/fd lookup), message-delivery (publish/consumeSync/
replay/listeners), stream-registry-lifecycle (onSessionOrphaned/Revived).

event-processor.ts (612 lines) → minimal extraction: publishWidgetEvent +
publishDocumentShow moved to session-broker-events.ts; delegation tracking
(#1075 block) extracted to DelegationTracker class in delegation-tracker.ts.

All original files preserved as thin re-export barrels — zero consumer
churn across 90 call sites and no import path changes in index.ts.
